refactor(week2-task1): reorganize book rendering logic
